webapi/webapi-mediaserver-tizen-tests/mediaserver/MediaServer_find_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2014 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Wang, Hongjuan <hongjuanx.wang@intel.com>

-->

<meta charset="utf-8">
<title>DLNA Test: MediaServer - find basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://01org.github.io/cloud-dleyna/mediaserver.html#widl-MediaServer-find-Promise-DOMString-containerId-DOMString-searchFilter-DOMString-sortMode-unsigned-long-count-unsigned-long-offset">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<div id='serverContainer'></div>
<div id='mediaObjectsContainer'></div>
<script>
  var t = async_test("Check if the MediaServer.find method can work well");
  var msga = navigator.mediaServer;
  msga.scanNetwork();
  
  function listMediaObjects(response) {
    t.step(function () {
      assert_equals(typeof response, "object", "type");
    });
    t.done();
  }
  
  function onFindFailure(error) {
    t.step(function () {
      assert_unreached("Error message: { " + error.message + " }");
    });
    t.done();
  }
  
  msga.onserverfound = function (event) {
    var container = document.getElementById('serverContainer');
    var searchField = document.createElement('input');
    var p = document.createElement('p')
    searchField.id = 'searchField';
    searchField.serverId = event.server.id;
    var selectElement = document.createElement('select');
    selectElement.id = 'searchParameter';
    event.server.searchAttrs.forEach(function(searchAttr) {
      var option = document.createElement('option');
      option.text = searchAttr;
      selectElement.add(option);
    });
    container.appendChild(p);
    p.appendChild(searchField);
    p.appendChild(selectElement);
    var mcontainer = document.getElementById('mediaObjectsContainer');
    while(mcontainer.hasChildNodes()){
      mcontainer.removeChild(mcontainer.lastChild);
    }
    container.serverId = searchField.serverId;
    var selectedIndex = selectElement.selectedIndex;
    var selectedParameter = selectElement.options[selectedIndex].text;
    searchField.value = "video";
    var searchValue = searchField.value;
    var query = selectedParameter + ' contains ' + '"' + searchValue + '"';
    event.server.find(event.server.root.id, query, '', 0, 0).then(listMediaObjects, onFindFailure);
  };
</script>
